Refactor units to avoid cyclic dependency.
